Migration file modified

+++ database/migrations/2025_06_04_062302_add_clock_date_to_clockings_table.php
@@ -8,15 +8,19 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('clockings', function (Blueprint $table) {
-            $table->date('clock_date')->after('time_out');
-        });
+        if (!Schema::hasColumn('clockings', 'clock_date')) {
+            Schema::table('clockings', function (Blueprint $table) {
+                $table->date('clock_date')->after('time_out');
+            });
+        }
     }
-    
+
     public function down(): void
     {
-        Schema::table('clockings', function (Blueprint $table) {
-            $table->dropColumn('clock_date');
-        });
+        if (Schema::hasColumn('clockings', 'clock_date')) {
+            Schema::table('clockings', function (Blueprint $table) {
+                $table->dropColumn('clock_date');
+            });
+        }
     }
 };
